+++ src/form-elements/FormElementSummary.tsx
@@ -5,7 +5,7 @@ import useFormSubmissionModel from '../hooks/useFormSubmissionModelContext'
 import useFlattenElements from '../hooks/useFlattenElementsContext'
 import { FormTypes } from '@oneblink/types'
 import { FormElementValueChangeHandler } from '../types/form'
-
+import getCorrectDateFromDateOnlyString from '../services/getCorrectDateFromDateOnlyString'
 type Props = {
   element: FormTypes.SummaryElement
   onChange: FormElementValueChangeHandler
@@ -123,8 +123,11 @@ function FormElementSummary({ element, onChange, value }: Props) {
         }
         case 'date': {
           if (typeof formElementValue !== 'string') return partialSummary
+          console.log(formElementValue)
           partialSummary.push(
-            localisationService.formatDate(new Date(formElementValue)),
+            localisationService.formatDate(
+              getCorrectDateFromDateOnlyString(formElementValue),
+            ),
           )
           break
         }